+++ Taller/HGibbs.h
@@ -1,9 +1,9 @@
 #ifndef HGIBBS_H_INCLUDED
 #define HGIBBS_H_INCLUDED
 
- long double hgibbs( long double r1[3],  long double r2[3],  long double r3[3],
-                                         long double MJD1, long double MJD2, long double MJD3,
-                                         long double vectVel[3],  long double angulos[2], char error[]);
+ double hgibbs( double r1[3],  double r2[3],  double r3[3],
+                                         double MJD1, double MJD2, double MJD3,
+                                         double vectVel[3],  double angulos[2], char error[]);
 
 
 #endif // HGIBBS_H_INCLUDED
