+++ Taller/MeanObliquity.h
@@ -1,6 +1,6 @@
 #ifndef MEANOBLIQUITY_H_INCLUDED
 #define MEANOBLIQUITY_H_INCLUDED
 
- long double MeanObliquity( long double Mjd_TT);
+ double MeanObliquity( double Mjd_TT);
 
 #endif // MEANOBLIQUITY_H_INCLUDED
