+++ Taller/Mjday.h
@@ -4,7 +4,7 @@
 #include <math.h>
 #include "MatlabUtilities.h"
 
- long double Mjday(int year, int month,int day, int hour, int m,  long double sec);
+ double Mjday(int year, int month,int day, int hour, int m,  double sec);
 
 
 #endif // MJDAY_H_INCLUDED
