+++ Taller/PrecMatrix.h
@@ -1,7 +1,7 @@
 #ifndef PRECMATRIX_H_INCLUDED
 #define PRECMATRIX_H_INCLUDED
 
-void PrecMatrix( long double ,  long double,  long double[3][3]);
+void PrecMatrix( double ,  double,  double[3][3]);
 
 
 #endif // PRECMATRIX_H_INCLUDED
