+++ Taller/R_x.h
@@ -1,6 +1,6 @@
 #ifndef R_X_H_INCLUDED
 #define R_X_H_INCLUDED
 
-void R_x( long double angle,  long double matrizRotada[3][3]);
+void R_x( double angle,  double matrizRotada[3][3]);
 
 #endif // R_X_H_INCLUDED
