+++ Taller/lambert_gooding.h
@@ -3,13 +3,13 @@
 
 #include <stdbool.h>
 
-void lambert_gooding( long double r1[3], long double r2[3],  long double tof,  long double mu, bool _way, int multi_revs,  long double v1[3],  long double v2[3]);
+void lambert_gooding( double r1[3], double r2[3],  double tof,  double mu, bool _way, int multi_revs,  double v1[3],  double v2[3]);
 
-int vlamb( long double gm, long double r1,  long double r2,  long double th,  long double tdelt, long double vri[2][1], long double vti[2][1], long double vrf[2][1], long double vtf[2][1] );
+int vlamb( double gm, double r1,  double r2,  double th,  double tdelt, double vri[2][1], double vti[2][1], double vrf[2][1], double vtf[2][1] );
 
-void tlamb( long double m, long double q, long double qsqfm1,  long double x,  long double n,  long double salida_t_dt_d2t_d3t[4]);
+void tlamb( double m, double q, double qsqfm1,  double x,  double n,  double salida_t_dt_d2t_d3t[4]);
 
-long double d8rt( long double x);
+double d8rt( double x);
 
-void xlamb ( long double m,  long double q,  long double qsqfm1,  long double tin,  long double salida_n_x_xpl[3]);
+void xlamb ( double m,  double q,  double qsqfm1,  double tin,  double salida_n_x_xpl[3]);
 #endif // LAMBERT_GOODING_H_INCLUDED
