+++ Taller/rv2coe.h
@@ -1,6 +1,6 @@
 #ifndef RV2COE_H_INCLUDED
 #define RV2COE_H_INCLUDED
 
-void rv2coe ( long double vectorPosicion[3],  long double vectorVelocidad[3],  long double vectorResultado[11]);
+void rv2coe ( double vectorPosicion[3],  double vectorVelocidad[3],  double vectorResultado[11]);
 
 #endif // RV2COE_H_INCLUDED
